fix: change to match new semicolon behaviour
This is a strange situation. I think the original test was passing in an unexpected way. ;busta was being added to the query parameters, but the check for semicolons looks at the r.URL.Path, which is always blank in these test cases. The previous http library behaviour was treating ; as a query delimiter, and so the test was passing because the parameter busta was unexpected, not because of the presence of the semicolon. The new behaviour drops any parameters that contain a ; , which was causing the tests to fail. To fix this, now the check for ; looks at r.URL.RawQuery, which does show ; . For services that treat ; as a valid URL delimiter, for example WFS, they must use the AllowQuerySemicolons handler wrapper, which converts all ; to & , which should get around this 'cache busting' check